Move JobDetails out of admin component
